requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.2.0
SQLAlchemy==2.0.35
PyMySQL==1.1.1
python-dotenv==1.0.1
Babel==2.16.0
cryptography==43.0.1

